+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.3.1 (October 27th, 2023)
+Typo
+
+## 2.3.0 (October 27th, 2023)
+This update allows you to use the Poseidon Customizer library through Zeus Core
+Fix bug with Fields translations
+Add a `default_term` argument to custom terms
+
 ## 2.2.0 (December 29th, 2020)
 Merry Christmas and Happy new year :)
 New Customizer, Control and Section integration